# Switch to heap-based deadline tracking