--- a/trio/_core/_run.py
+++ b/trio/_core/_run.py
@@ -20,6 +20,7 @@ from time import perf_counter
 from sniffio import current_async_library_cvar
 
 import attr
+from heapq import heapify, heappop, heappush
 from sortedcontainers import SortedDict
 from outcome import Error, Value, capture
 
@@ -121,6 +122,55 @@ class IdlePrimedTypes(enum.Enum):
 ################################################################
 # CancelScope and friends
 ################################################################
+
+
+@attr.s(eq=False, slots=True)
+class Deadlines:
+    """A container of deadlined cancel scopes.
+
+    Only contains scopes with non-infinite deadlines that are currently
+    attached to at least one task.
+
+    """
+
+    # Heap of (deadline, id(CancelScope), CancelScope)
+    _heap = attr.ib(factory=list)
+    # Count of active deadlines (those that haven't been changed)
+    _active = attr.ib(default=0)
+
+    def add(self, deadline, cancel_scope):
+        heappush(self._heap, (deadline, id(cancel_scope), cancel_scope))
+        self._active += 1
+
+    def remove(self, deadline, cancel_scope):
+        self._active -= 1
+
+    def next_deadline(self):
+        while self._heap:
+            deadline, _, cancel_scope = self._heap[0]
+            if deadline == cancel_scope._registered_deadline:
+                return deadline
+            else:
+                # This entry is stale; discard it and try again
+                heappop(self._heap)
+        return inf
+
+    def expire(self, now):
+        did_something = False
+        while self._heap and self._heap[0][0] <= now:
+            deadline, _, cancel_scope = heappop(self._heap)
+            if deadline == cancel_scope._registered_deadline:
+                did_something = True
+                # This implicitly calls self.remove(), so we don't need to
+                # decrement _active here
+                cancel_scope.cancel()
+        # If we've accumulated too many stale entries, then prune the heap to
+        # keep it under control. (We only do this occasionally in a batch, to
+        # keep the amortized cost down)
+        if len(self._heap) > self._active * 2 + 1000:
+            self._heap = [t for t in self._heap if t[0] == t[2]._registered_deadline]
+            heapify(self._heap)
+        return did_something
 
 
 @attr.s(eq=False, slots=True)
@@ -518,13 +568,13 @@ class CancelScope(metaclass=Final):
                 self._registered_deadline = new
                 runner = GLOBAL_RUN_CONTEXT.runner
                 if runner.is_guest:
-                    old_next_deadline = runner.next_deadline()
+                    old_next_deadline = runner.deadlines.next_deadline()
                 if old != inf:
-                    del runner.deadlines[old, id(self)]
+                    runner.deadlines.remove(old, self)
                 if new != inf:
-                    runner.deadlines[new, id(self)] = self
+                    runner.deadlines.add(new, self)
                 if runner.is_guest:
-                    new_next_deadline = runner.next_deadline()
+                    new_next_deadline = runner.deadlines.next_deadline()
                     if old_next_deadline != new_next_deadline:
                         runner.force_guest_tick_asap()
 
@@ -1201,10 +1251,7 @@ class Runner:
     runq = attr.ib(factory=deque)
     tasks = attr.ib(factory=set)
 
-    # {(deadline, id(CancelScope)): CancelScope}
-    # only contains scopes with non-infinite deadlines that are currently
-    # attached to at least one task
-    deadlines = attr.ib(factory=SortedDict)
+    deadlines = attr.ib(factory=Deadlines)
 
     init_task = attr.ib(default=None)
     system_nursery = attr.ib(default=None)
@@ -1236,14 +1283,6 @@ class Runner:
         # This is where KI protection gets disabled, so we do it last
         self.ki_manager.close()
 
-    def next_deadline(self):
-        try:
-            (next_deadline, _), _ = self.deadlines.peekitem(0)
-        except IndexError:
-            return inf
-        else:
-            return next_deadline
-
     @_public
     def current_statistics(self):
         """Returns an object containing run-loop-level debugging information.
@@ -1268,7 +1307,7 @@ class Runner:
           other attributes vary between backends.
 
         """
-        seconds_to_next_deadline = self.next_deadline() - self.current_time()
+        seconds_to_next_deadline = self.deadlines.next_deadline() - self.current_time()
         return _RunStatistics(
             tasks_living=len(self.tasks),
             tasks_runnable=len(self.runq),
@@ -2009,11 +2048,9 @@ def unrolled_run(runner, async_fn, args, host_uses_signal_set_wakeup_fd=False):
         while runner.tasks:
             if runner.runq:
                 timeout = 0
-            elif runner.deadlines:
-                deadline, _ = runner.deadlines.keys()[0]
-                timeout = runner.clock.deadline_to_sleep_time(deadline)
             else:
-                timeout = _MAX_TIMEOUT
+                deadline = runner.deadlines.next_deadline()
+                timeout = runner.clock.deadline_to_sleep_time(deadline)
             timeout = min(max(0, timeout), _MAX_TIMEOUT)
 
             idle_primed = None
@@ -2042,14 +2079,8 @@ def unrolled_run(runner, async_fn, args, host_uses_signal_set_wakeup_fd=False):
 
             # Process cancellations due to deadline expiry
             now = runner.clock.current_time()
-            while runner.deadlines:
-                (deadline, _), cancel_scope = runner.deadlines.peekitem(0)
-                if deadline <= now:
-                    # This removes the given scope from runner.deadlines:
-                    cancel_scope.cancel()
-                    idle_primed = None
-                else:
-                    break
+            if runner.deadlines.expire(now):
+                idle_primed = None
 
             # idle_primed != None means: if the IO wait hit the timeout, and
             # still nothing is happening, then we should start waking up