@njsmith 2020-06-20
So there is one potential flaw I noticed in this; I'm not sure if we should care about it or not. Consider some pathological code that repeatedly switches a deadline back and forth between two distinct values:

```python
with CancelScope() as cscope:
    for _ in range(100000):
        cscope.deadline = 100
        cscope.deadline = 101
```

This creates tons and tons of entries in our heap, most of them stale. Eventually this will trigger a heap prune. But the prune will end up leaving a bunch of the stale entries in-place, because they all have a `deadline` that matches the current `cscope._registered_deadline`. And then we enter a pathological state, where we keep attempting to prune on every scheduler tick, while accomplishing nothing, creating a nasty slowdown.

Now that I write this, it occurs to me that there's a pretty trivial solution: have the prune operation keep a set of cancel scope objects it's seen, and drop duplicates. That adds a little bit of overhead to the prune operation, but it shouldn't be too bad.