@njsmith 2020-06-20
> Now that I write this, it occurs to me that there's a pretty trivial solution: have the prune operation keep a set of cancel scope objects it's seen, and drop duplicates. That adds a little bit of overhead to the prune operation, but it shouldn't be too bad.

Done.